Implementação Inicial do front - FALTA CONTROLE DE ROTAS PRO BACK+detalhes

+++ api/src/controllers/ingredients.go
@@ -9,12 +9,11 @@ import (
 	"api/src/config"
 	"encoding/json"
 	"github.com/gorilla/mux"
-	"time"
+	"strconv"
 )
 
 //Insere Ingredient no db
 func CreateIngredient(w http.ResponseWriter, r *http.Request) {
-	//w.Write([]byte("Criando ingredient_id!"))
 	corpoRequest, erro := ioutil.ReadAll(r.Body)
 	if erro != nil {
 		log.Fatal(erro)
@@ -31,11 +30,19 @@ func CreateIngredient(w http.ResponseWriter, r *http.Request) {
 		log.Fatal(erro)
 	}
 
+	if  len(ingredient.Nome)==0 || len(ingredient.Description)==0 {
+		w.WriteHeader(400)
+		w.Write([]byte("Erro ao criar Ingredient, campo(s) nulos"))
+		fmt.Println("Erro ao criar Ingredient, campo(s) nulos")
+		return
+    }
+
 	query := "INSERT INTO ingredients (nome, description) VALUES ('"+ingredient.Nome+"','"+ingredient.Description+"') RETURNING id_ingredients"
 
 	sqlStatement, err := db.Query(query)
 	if err != nil {
 		w.WriteHeader(400)
+		w.Write([]byte("Erro ao criar Ingredient"))
         panic(err.Error())
     }
 
@@ -45,10 +52,15 @@ func CreateIngredient(w http.ResponseWriter, r *http.Request) {
         if err != nil {
 			panic(err.Error())
 		}
-		w.Write([]byte(id))
-		fmt.Println("Usuário Criado com id: ",id)
+		aux_ID,err_ID := strconv.Atoi(id)
+		if err_ID != nil {
+			panic(err.Error())
+		}
+		ingredient.ID = uint64(aux_ID)
+		json.NewEncoder(w).Encode(ingredient)
+		fmt.Println("Ingredient Criado com id: ",id)
+		return
 	}
-	
 
 	defer db.Close()
 }
@@ -62,26 +74,82 @@ func GetAllIngredients(w http.ResponseWriter, r *http.Request) {
         panic(err.Error())
     }
 
-	query := "SELECT * from ingredients"
-
+	//query := "SELECT * from ingredients"
+	query := "select array_to_json(array_agg(row_to_json(ingredients_alias))) from (select id_ingredients as \"id\",nome as \"nome\", description as \"description\", datacriacao as \"dataCriacao\" from ingredients) ingredients_alias"
 	sqlStatement, err := db.Query(query)
 	if err != nil {
         panic(err.Error())
     }
 
-	aux := "["
+	//Funcional
+	var aux string
 	for sqlStatement.Next() {
-		var ingredient models.Ingredient
-        err = sqlStatement.Scan(&ingredient.ID, &ingredient.Nome, &ingredient.Description, &ingredient.DataCriacao)
+        err = sqlStatement.Scan(&aux)
         if err != nil {
 			panic(err.Error())
 		}
-		str := fmt.Sprintf("{\"Id\":\"%d\",\"Nome\":\"%s\",\"Description\":\"%s\", \"DataCriacao\":\"%s\"},", ingredient.ID, ingredient.Nome, ingredient.Description, ingredient.DataCriacao.Format(time.RFC822))
-		aux = aux + str
     }
-	aux = trimLastChar(aux)
-	aux = aux + "]"
 	w.Write([]byte(aux))
+
+	//Também Funcional, porém podendo mudar o padrão time aplicado
+	//aux := "["
+	//for sqlStatement.Next() {
+	//	var ingredient models.Ingredient
+    //    err = sqlStatement.Scan(&ingredient.ID, &ingredient.Nome, &ingredient.Description, &ingredient.DataCriacao)
+    //    if err != nil {
+	//		panic(err.Error())
+	//	}
+	//	str := fmt.Sprintf("{\"Id\":\"%d\",\"Nome\":\"%s\",\"Description\":\"%s\", \"DataCriacao\":\"%s\"},", ingredient.ID, ingredient.Nome, ingredient.Description, ingredient.DataCriacao.Format(time.RFC822))
+	//	aux = aux + str
+    //}
+	//aux = trimLastChar(aux)
+	//aux = aux + "]"
+	//w.Write([]byte(aux))
+
+	fmt.Println("Listando todos os ingredients")
+
+	defer db.Close()
+}
+
+func GetAllIngredientsIds(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "aplication/json")
+
+	db, err := config.Conectar()
+	if db == nil {
+        panic(err.Error())
+    }
+
+	//query := "SELECT * from ingredients"
+	query := "select array_to_json(array_agg(row_to_json(ingredients_alias))) from (select id_ingredients as \"id\" from ingredients) ingredients_alias"
+	sqlStatement, err := db.Query(query)
+	if err != nil {
+        panic(err.Error())
+    }
+
+	//Funcional
+	var aux string
+	for sqlStatement.Next() {
+        err = sqlStatement.Scan(&aux)
+        if err != nil {
+			panic(err.Error())
+		}
+    }
+	w.Write([]byte(aux))
+
+	//Também Funcional, porém podendo mudar o padrão time aplicado
+	//aux := "["
+	//for sqlStatement.Next() {
+	//	var ingredient models.Ingredient
+    //    err = sqlStatement.Scan(&ingredient.ID, &ingredient.Nome, &ingredient.Description, &ingredient.DataCriacao)
+    //    if err != nil {
+	//		panic(err.Error())
+	//	}
+	//	str := fmt.Sprintf("{\"Id\":\"%d\",\"Nome\":\"%s\",\"Description\":\"%s\", \"DataCriacao\":\"%s\"},", ingredient.ID, ingredient.Nome, ingredient.Description, ingredient.DataCriacao.Format(time.RFC822))
+	//	aux = aux + str
+    //}
+	//aux = trimLastChar(aux)
+	//aux = aux + "]"
+	//w.Write([]byte(aux))
 
 	fmt.Println("Listando todos os ingredients")
 
@@ -104,7 +172,7 @@ func GetIngredientById(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
         panic(err.Error())
     }
-	w.WriteHeader(200)
+	
 	for sqlStatement.Next() {
 
         var ingredient models.Ingredient
@@ -114,6 +182,12 @@ func GetIngredientById(w http.ResponseWriter, r *http.Request) {
 			panic(err.Error())
 		}
 		
+		//if (models.Ingredient{}) == ingredient {
+		//	w.WriteHeader(404)
+		//	w.Write([]byte("Ingredient não encontrado"))
+		//	fmt.Println("ingredient_id:", params["ingredients_id"]," NÃO foi encontrado")
+		//	return
+		//}
 		json.NewEncoder(w).Encode(ingredient)
 		fmt.Println("ingredient_id:", params["ingredients_id"]," foi encontrado")
 		return
@@ -144,6 +218,19 @@ func UpdateIngredientById(w http.ResponseWriter, r *http.Request) {
 		log.Fatal(erro)
 	}
 
+	if  len(ingredient.Nome)==0 || len(ingredient.Description)==0 {
+		w.WriteHeader(400)
+		w.Write([]byte("Erro ao atualizar Ingredient, campo(s) nulos"))
+		fmt.Println("Erro ao atualizar Ingredient, campo(s) nulos")
+		return
+    }
+
+	if confere_id((params["ingredients_id"])) {
+		fmt.Println("ingredient_id:", params["ingredients_id"]," não foi encontrado")
+		w.WriteHeader(404)
+		return
+	}
+	
 	//query := "INSERT INTO ingredients (nome, description) VALUES ('"+ingredient.Nome+"','"+ingredient.Description+"')"
 	query := "UPDATE ingredients set nome = '"+ingredient.Nome+"', description = '"+ingredient.Description+"' where id_ingredients = "+params["ingredients_id"]
 
@@ -172,6 +259,12 @@ func DeleteIngredientById(w http.ResponseWriter, r *http.Request) {
         panic(err.Error())
     }
 
+	if confere_id((params["ingredients_id"])) {
+		fmt.Println("ingredient_id:", params["ingredients_id"]," não foi encontrado")
+		w.WriteHeader(404)
+		return
+	}
+
 	query := "Delete from ingredients where id_ingredients = " + params["ingredients_id"]
 
 	sqlStatement, err := db.Query(query)
@@ -185,6 +278,34 @@ func DeleteIngredientById(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(410)
 	fmt.Println("ingredient_id:", params["ingredients_id"]," foi DELETADO")
 	defer db.Close()
+}
+
+func confere_id(id string) bool{
+	db, err := config.Conectar()
+	if db == nil {
+        panic(err.Error())
+    }
+
+	query := "SELECT id_ingredients from ingredients"
+
+	sqlStatement, err := db.Query(query)
+	if err != nil {
+        panic(err.Error())
+    }
+
+	for sqlStatement.Next() {	
+		var tmp string
+		err = sqlStatement.Scan(&tmp)
+        if err != nil {
+			panic(err.Error())
+		}
+
+		if tmp == id {
+			return false
+		}
+    }
+	defer db.Close()
+	return true
 }
 
 func trimLastChar(s string) string {
